trading_engine/requirements.txt
ccxt==4.2.15
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
websockets==11.0.3
loguru==0.7.2
pydantic==1.10.15
# Backtesting
backtrader==1.9.78.123
vectorbt==0.26.0
# Technical Analysis
ta==0.9.0
# Data visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0
# Database (optional)
sqlalchemy==2.0.23
# Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
